Return values for all cpus from lookup on percpu maps
With percpu maps, one data structure exists for every cpu. When a lookup
is performed on the map, libbpf copies one value per cpu to the output.
Previously, only the first value was then returned to the users of this
crate. In addition, some memory was overwritten because the reserved
area for the values was too small.

This commit reserves space for all values from a percpu map and returns
all values to the users of this crate.